Implement TensorflowTrotterHamiltonian
